+++ outputs.tf
@@ -4,7 +4,7 @@ locals {
   # Just using the name here will not create this dependency as the name is known
   # in the planning phase of terraform.
   # node pools are created using the name only (no self_link, no id is working)
-  name_from_id = element(reverse(split("/", google_container_cluster.cluster[0].id)), 0)
+  name_from_id = try(element(reverse(split("/", google_container_cluster.cluster[0].id)), 0), null)
 }
 
 # ----------------------------------------------------------------------------------------------------------------------