+++ test/unit-disabled/main.tf
@@ -1,0 +1,45 @@
+# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+# EMPTY FEATURES (DISABLED) UNIT TEST
+# This module tests an empty set of features.
+# The purpose is to verify no resources are created when the module is disabled.
+# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+variable "project" {
+  type        = string
+  description = "(Required) The ID of the project in which the resource belongs."
+}
+
+terraform {
+  required_providers {
+    google = {
+      source  = "hashicorp/google"
+      version = "~> 4.0"
+    }
+  }
+}
+
+provider "google" {
+  project = var.project
+}
+
+# DO NOT RENAME MODULE NAME
+module "test" {
+  source = "../.."
+
+  module_enabled = false
+
+  # add all required arguments
+  name       = "gke-unit-disabled"
+  network    = "disabled"
+  subnetwork = "disabled"
+
+  # add all optional arguments that create additional resources
+
+  # add only required arguments and no optional arguments
+}
+
+# outputs generate non-idempotent terraform plans so we disable them for now unless we need them.
+# output "all" {
+#   description = "All outputs of the module."
+#   value       = module.test
+# }